Move traits to prelude
